split bluespec
